Create Iterator.hpp
This is my current progress on the iterator.hpp file. This file is still incomplete as I'm currently working on it. This provides a set of abstract base classes that force the user to implement required functions when making an iterator class. This will allow custom iterator classes to universally work with global iterator functions such as std::accumulate