Update Type_Traits.hpp
I moved some meta-functions over to Utility.hpp and moved some other meta-functions to this file to change their implementations. I added a few more meta-functions as well. This will most likely be the last update to this file until I can test the commented out implementations which I may switch to if their prove to be viable. I may still add more meta-functions, but at this point it is unlikely to happen for a while until I start developing the actual data structure classes.