Create Concepts.hpp
This is my original file to implement concept checks, but I originally remade all of the current libstdc++(c++20) concepts using SFINAE meta-functions. This isn't as streamlined or efficient as the Concep_Checks.hpp implementations which are from c++14. I might still use this file, so I'm still adding it to this repo. I will most likely just remove this file completely though.