Create UtilityBase.hpp
This file is a consolidation of a bunch of headers from libstdc++. Some of these headers can toggle between 2 implementations from different c++ standards(like c++14 and c++20). This file serves as the backbone for this whole library along with Type_Traits.hpp, so it just has general required utility headers unlike Utility.hpp which is more specialized for containers. There are a lot of different headers implemented in this file, so there is a chance I will need to split some headers off into their own files if this ends up causing include issues.