Create Array.hpp
This is the implementation of std::array from libstdc++(gcc-4.9.0). This file is used as a base for some of the other sequence container classes, so this file will be kept in the util/ folder. This class just encapsulates an array to make using arrays much easier instead of having to manually mess around with pointers yourself. This class meets all of the requirements for a container, so it is the first actual container class.