Create Debug_Array.hpp
This is a parallel implementation of std::array which gets compiled instead of the main implementation when the debug mode macro is defined. This file is kept in the util/ folder since the main Array.hpp file is also in the util/ folder. This is done since the arduino IDE can't include files like in traditional c++, so you cannot include through the parent folder since thats where the debug/ folder is. I will test it more later to see if this being a library will make a difference.
Ideally doing #include "/../debug/Debug_Array.hpp" should work, but it doesn't in the arduino IDE due to how the compile files are stored which is not in the same file structure as the actual sketch folder(which is kinda annoying).