Create Tuple.hpp
This file is for the implementation of std::tuple from libstdc++(gcc-4.9.0). Since this class is used as a base for some container classes this file will also be in the util/ folder along with std::pair and std::array. The std::tuple class is a fixed size data structure that holds a set amount of elements like an array, but each element can be any type(like int, string, char, object). This is the most complicated data structure to implement in c++ STL, so this is pretty much an exact copy of the libstdc++ source code with slight preprocessor and stylizing/format changes.