Create Vector.hpp
This file is the main declaration/definition of the std::vector class from libstdc++(gcc-4.9.0). There are some large functions that are defined separately in the Vector.hh file which gets included at the end of this file. This class also has a debug mode version which is included from the debug/ folder. This class functions like an array, but it's size is dynamic. This container will automatically increase it's capacity and reallocate its memory to the new larger chunk of memory when the capacity is exceeded.