Create Vector.hh
This is an implementation of the <bits/vector.tcc> header from libstdc++(gcc-4.9.0). This file just defines the large functions from std::vector which are declared but not defined in the original class declaration. I might just merge this with the main Vector.hpp file since there most likely wont be any point for the export template functionality. This file also uses the ".hh" file extension instead of ".tcc" since the Arduino IDE doesn't support ".tcc" files, and ".tpp" files aren't recognized properly(they also dont have the visual theme in the IDE, so everything is just white and difficult to read). This file is included at the end of the main Vector.hpp file. Under no circumstance should this file be included directly anywhere else except Vector.hpp. Also alot of the functions in this file already do a bunch of checks, so they are not overloaded in the debug mode version of std::vector as there are no additional checks to be made.