Create Debug_Vector.hpp
This is the debug mode implementation for the std::vector class. This just creates a std::vector class wrapper that gets used instead of the main vector class when debug mode is enabled. This wrapper overloads certain members in the class to do more checks to check for any code that can cause bugs. I might end up moving this file along with everything in the debug/ folder over to the src/ folder with the main implementations since this file needs to include from the parent directory, and it needs to include through the parent folder to get files from the util/folder.
If the parent include statements don't work even with libraries in the Arduino IDE then I will get rid of the debug folder. I will most likely be a while before I can get to properly test this file structure since in the Arduino sketch I'm using to develop this library everything is all just in one folder so that I can actually see the files in the Arduino IDE.