Update README.md
Updating the readme to reflect the new direction the development is going in. I removed a bunch of stuff referring to how I was  previously implementing everything. Now mostly everything is only a slightly modified version of the libstdc++(gcc-4.9.0) implementations. I also updated the to-do list to reflect my current progress and plans. I also added a section about the customizable reallocation I plan on adding which is to make this library more viable for lower memory boards.